% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_pets.R
\name{num_pets}
\alias{num_pets}
\title{Number of pets you have}
\usage{
num_pets(name, number, animal)
}
\arguments{
\item{name}{a character containing the input name}

\item{number}{a number containing the number of pets the person has}

\item{animal}{a character containing the type of pet}
}
\value{

}
\description{
Contains one function, num_pets, that prints a statement about the number of pets a user has.
}
