% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_blue_white.R
\name{theme_blue_white}
\alias{theme_blue_white}
\title{Blue and white theme}
\usage{
theme_blue_white()
}
\value{

}
\description{
Blue background, white panel background with gray grid lines, purple axis labels and teal axis titles.
}
