% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel.R
\name{travel}
\alias{travel}
\title{Favorite travels}
\usage{
travel(name, place)
}
\arguments{
\item{name}{a string containing the user's name}

\item{place}{a string containing the user's favorite place they've traveled to.}
}
\value{

}
\description{
Returns a statement of where a user's favorite place they've traveled to is.
}
