% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekend_plans.R
\name{weekend_plans}
\alias{weekend_plans}
\title{Weekend Plans}
\usage{
weekend_plans(day, activity)
}
\arguments{
\item{day}{a string containing the day that the user plans on doing the activity}

\item{activity}{a string containing the activity the user plans on doing}
}
\value{

}
\description{
This function returns what activity the user plans on doing on a weekend day.
}
